107 Refactor edit game + add elements for toggling active games
