+++ src/types/gameDto.ts
@@ -1,5 +1,7 @@
 import { GameDescription } from './gameResponse'
 
+export type ActiveGameType = 'all' | 'custom' | 'none'
+
 type Game = {
     id: number
     created_at: string
